Implement f64

+++ src/activation.rs
@@ -1,8 +1,8 @@
-use std::f32::consts::E;
+use std::f64::consts::E;
 
-pub type Activation = fn(f32) -> f32;
+pub type Activation = fn(f64) -> f64;
 
-pub fn binary_step(x: f32) -> f32 {
+pub fn binary_step(x: f64) -> f64 {
     if x >= 0.0 {
         1.0
     } else {
@@ -18,11 +18,11 @@ pub fn linear(x: f64) -> f64 {
     identity(x)
 }
 
-pub fn logistic(x: f32) -> f32 {
+pub fn logistic(x: f64) -> f64 {
     1.0 / (1.0 + E.powf(-x))
 }
 
-pub fn sigmoid(x: f32) -> f32 {
+pub fn sigmoid(x: f64) -> f64 {
     logistic(x)
 }
